Add type checking for stats computation and split up definitions.
Also make most stats rounding 1 to reduce the significant figures and
information overload.
